Evaluate When

Evaluate expressions against a Spark DataFrame (or data.frame) if a predicate evaluates to TRUE. Note that these functions are subtly different from the scoped variants of dplyr functions in that they can evaluate any predicate.

arrange_when(.data, .eval, ...)

distinct_when(.data, .eval, ...)

filter_when(.data, .eval, ...)

group_by_when(.data, .eval, ...)

mutate_when(.data, .eval, ...)

select_when(.data, .eval, ...)

summarise_when(.data, .eval, ...)

transmute_when(.data, .eval, ...)

Arguments

.data
A Spark DataFrame or a data.frame.

.eval
A formula(). The LHS is the predicate to evaluate. When it evaluates to TRUE, the RHS will be performed in the context of the function. The RHS can be in one of several forms:

An expression, e.g. x * y

A named expression, e.g. c(z = x * y)

A vector or list of the above.

Note: named expressions must be wrapped in c() or list().

...
Additional parameters to pass to the dplyr function.

Value

A Spark DataFrame or a data.frame depending on the input, .data.

Examples

# Unname expressions will return columns with the expression as the name
mtcars %>% mutate_when("cyl" %in% colnames(mtcars) ~ mpg * 2)

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb mpg * 2
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4    42.0
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4    42.0
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1    45.6
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1    42.8
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2    37.4
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1    36.2
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4    28.6
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2    48.8
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2    45.6
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4    38.4
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4    35.6
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3    32.8
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3    34.6
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3    30.4
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4    20.8
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4    20.8
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4    29.4
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1    64.8
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2    60.8
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1    67.8
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1    43.0
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2    31.0
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2    30.4
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4    26.6
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2    38.4
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1    54.6
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2    52.0
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2    60.8
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4    31.6
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6    39.4
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8    30.0
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2    42.8

# But we can name them easily enough
mtcars %>% mutate_when("cyl" %in% colnames(mtcars) ~ c(mpg2 = mpg * 2))

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb mpg2
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 42.0
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 42.0
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 45.6
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 42.8
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 37.4
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 36.2
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 28.6
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 48.8
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 45.6
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 38.4
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 35.6
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 32.8
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 34.6
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 30.4
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 20.8
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 20.8
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 29.4
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 64.8
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 60.8
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 67.8
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 43.0
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 31.0
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 30.4
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 26.6
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 38.4
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 54.6
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 52.0
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 60.8
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 31.6
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 39.4
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 30.0
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 42.8

# We can evaluate multiple expressions
mtcars %>%
  mutate_when("cyl" %in% colnames(mtcars) ~ c(mpg2 = mpg * 2, mpg4 = mpg * 2))

#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb mpg2 mpg4
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 42.0 42.0
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 42.0 42.0
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 45.6 45.6
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 42.8 42.8
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 37.4 37.4
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 36.2 36.2
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 28.6 28.6
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 48.8 48.8
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 45.6 45.6
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 38.4 38.4
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 35.6 35.6
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 32.8 32.8
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 34.6 34.6
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 30.4 30.4
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 20.8 20.8
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 20.8 20.8
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 29.4 29.4
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 64.8 64.8
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 60.8 60.8
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 67.8 67.8
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 43.0 43.0
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 31.0 31.0
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 30.4 30.4
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 26.6 26.6
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 38.4 38.4
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 54.6 54.6
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 52.0 52.0
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 60.8 60.8
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 31.6 31.6
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 39.4 39.4
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 30.0 30.0
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 42.8 42.8

# We can still use functionality such as tidy-select
mtcars %>%
  select_when("cyl" %in% colnames(mtcars) ~ c(Cylinders = cyl, everything()))

#>                     Cylinders  mpg  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4                   6 21.0 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag               6 21.0 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710                  4 22.8 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive              6 21.4 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout           8 18.7 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant                     6 18.1 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360                  8 14.3 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D                   4 24.4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230                    4 22.8 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280                    6 19.2 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C                   6 17.8 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE                  8 16.4 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL                  8 17.3 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC                 8 15.2 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood          8 10.4 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental         8 10.4 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial           8 14.7 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128                    4 32.4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic                 4 30.4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla              4 33.9  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona               4 21.5 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger            8 15.5 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin                 8 15.2 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28                  8 13.3 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird            8 19.2 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9                   4 27.3  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2               4 26.0 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa                4 30.4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L              8 15.8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino                6 19.7 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora               8 15.0 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E                  4 21.4 121.0 109 4.11 2.780 18.60  1  1    4    2